chore: util 클래스 util 패키지 아래로 변경
